update code per PR review comments

+++ 0.download_data/scripts/update_file_structure.py
@@ -38,21 +38,10 @@ image_extensions = {".tif", ".tiff"}
 # In[3]:
 
 
-# Iterate over all folders at the first level inside the parent directory
-for first_level_folder in parent_dir.iterdir():
-    if first_level_folder.is_dir():
-        # Go into the first folder
-        for second_level_folder in first_level_folder.iterdir():
-            if second_level_folder.is_dir():
-                # Go into the second folder
-                for third_level_folder in second_level_folder.iterdir():
-                    if third_level_folder.is_dir():
-                        # Get all images inside the third-level folder
-                        for image_file in third_level_folder.iterdir():
-                            if image_file.suffix.lower() in image_extensions:
-                                # Copy each image to the NF0014 folder
-                                destination = nf0014_dir / image_file.name
-                                shutil.copy2(image_file, destination)  # copy2 preserves metadata
+for image_file in parent_dir.rglob('*/*/*/*'): # goes into third nested folder
+    if image_file.is_file() and image_file.suffix.lower() in image_extensions:
+        destination = nf0014_dir / image_file.name
+        shutil.copy2(image_file, destination)  # copy2 preserves metadata
 
 print("All images have been copied to the NF0014 folder!")
